app is working, still tests to fix

+++ src/Controller/TaskController.php
@@ -4,6 +4,8 @@ namespace App\Controller;
 
 use App\Entity\Task;
 use App\Form\TaskType;
+use Doctrine\Persistence\ManagerRegistry;
+use http\Exception\RuntimeException;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -11,6 +13,13 @@ use Symfony\Component\HttpFoundation\Response;
 
 class TaskController extends AbstractController
 {
+    private $doctrine;
+
+    public function __construct(ManagerRegistry $doctrine)
+    {
+        $this->doctrine = $doctrine;
+    }
+
     /**
      * @Route("/tasks", name="task_list")
      */
@@ -18,7 +27,7 @@ class TaskController extends AbstractController
     {
         return $this->render(
             'task/list.html.twig',
-            ['tasks' => $this->getDoctrine()->getRepository(Task::class)->findAll()]
+            ['tasks' => $this->doctrine->getRepository(Task::class)->findAll()]
         );
     }
 
@@ -33,7 +42,7 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
+            $em = $this->doctrine->getManager();
 
             $em->persist($task);
             $em->flush();
@@ -60,7 +69,7 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $this->doctrine->getManager()->flush();
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
 
@@ -76,10 +85,17 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/toggle", name="task_toggle")
      */
-    public function toggleTaskAction(Task $task)
+    public function toggleTaskAction(int $id)
     {
+        $task = $this->doctrine->getRepository(Task::class)->find($id);
+
+        if (!$task) {
+            throw new RuntimeException('Task not found');
+        }
+
         $task->toggle(!$task->isDone());
-        $this->getDoctrine()->getManager()->flush();
+        $this->doctrine->getManager()->persist($task);
+        $this->doctrine->getManager()->flush();
 
         $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
 
@@ -89,9 +105,15 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/delete", name="task_delete")
      */
-    public function deleteTaskAction(Task $task)
+    public function deleteTaskAction(int $id)
     {
-        $em = $this->getDoctrine()->getManager();
+        $task = $this->doctrine->getRepository(Task::class)->find($id);
+
+        if (!$task) {
+            throw new RuntimeException('Task not found');
+        }
+
+        $em = $this->doctrine->getManager();
         $em->remove($task);
         $em->flush();
